policy: clean-up
