+++ src/tools/genpolicy/src/pod.rs
@@ -11,7 +11,6 @@ use crate::infra;
 use crate::obj_meta;
 use crate::policy;
 use crate::registry;
-use crate::utils;
 use crate::volumes;
 use crate::yaml;
 
@@ -249,7 +248,7 @@ impl EnvVar {
 #[async_trait]
 impl yaml::K8sObject for Pod {
     fn get_metadata_name(&self) -> String {
-        utils::get_metadata_name(&self.metadata)
+        self.metadata.get_name()
     }
 
     fn get_host_name(&self) -> String {
@@ -262,15 +261,15 @@ impl yaml::K8sObject for Pod {
     }
 
     fn get_namespace(&self) -> String {
-        utils::get_metadata_namespace(&self.metadata)
+        self.metadata.get_namespace()
     }
 
     fn add_policy_annotation(&mut self, encoded_policy: &str) {
-        utils::add_policy_annotation(&mut self.metadata, encoded_policy)
+        self.metadata.add_policy_annotation(encoded_policy)
     }
 
     async fn get_registry_containers(&self) -> Result<Vec<registry::Container>> {
-        utils::get_registry_containers(&self.spec.containers).await
+        registry::get_registry_containers(&self.spec.containers).await
     }
 
     fn get_policy_data(
